Moved Send + Sync test to a new tests module in src/errors.rs
